Re-arrange code
