+++ stag_lib/KDE/data.cpp
@@ -2,6 +2,7 @@
    This file is provided as part of the STAG library and released under the GPL
    license.
 */
+// Standard C++ libraries
 #include <iostream>
 #include <fstream>
 #include <random>
@@ -9,14 +10,14 @@
 #include <deque>
 #include <future>
 
+// Additional libraries
+#include "multithreading/ctpl_stl.h"
+
+// STAG modules
 #include "data.h"
 #include "utility.h"
 #include "Graph/random.h"
 #include "KDE/kde.h"
-#include "multithreading/ctpl_stl.h"
-#include "indicators/indicators.hpp"
-
-#define KDE_TREE_CUTOFF 5000
 
 /*
  * Used to disable compiler warning for unused variable.
@@ -188,355 +189,4 @@ DenseMat stag::load_matrix(std::string& filename) {
   data.conservativeResize(num_rows, Eigen::NoChange_t::NoChange);
 
   return data;
-}
-
-//------------------------------------------------------------------------------
-// Constructing approximate similarity graph.
-//------------------------------------------------------------------------------
-class KDETreeEntry {
-public:
-  KDETreeEntry(DenseMat* data, StagReal a, StagInt min_id, StagInt max_id, StagInt dep, StagInt dep_cutoff)
-    : sampling_dist(0.0, 1.0)
-  {
-     min_idx = min_id;
-     max_idx = max_id;
-     n_node = max_idx - min_idx;
-     depth = dep;
-     depth_cutoff = dep_cutoff;
-
-     if (max_idx - min_idx >= KDE_TREE_CUTOFF) {
-       below_cutoff = false;
-
-       // We only initialise an estimator at this node if we are below the depth
-       // depth cutoff. Above the depth cutoff and we just use the children to estimate.
-       if (depth > depth_cutoff) initialise_estimator(data, a);
-     } else {
-       below_cutoff = true;
-       initialise_exact(data, a);
-     }
-
-     if (!below_cutoff) {
-       StagInt midpoint = (min_idx + max_idx) / 2;
-       assert(midpoint >= min_idx);
-       assert(midpoint < max_idx);
-       left_child = new KDETreeEntry(data, a, min_idx, midpoint, depth + 1, depth_cutoff);
-       right_child = new KDETreeEntry(data, a, midpoint + 1, max_idx, depth + 1, depth_cutoff);
-     }
-  }
-
-  void initialise_estimator(DenseMat* data, StagReal a) {
-    StagInt K1 = ceil(1 * log((StagReal) n_node));
-    StagReal K2_constant = 0.1;
-    StagReal min_mu = 1.0 / (StagReal) n_node;
-    StagInt offset = 0;
-    this_estimator = stag::CKNSGaussianKDE(data, a, min_mu, K1, K2_constant,
-                                           offset, min_idx, max_idx + 1);
-  }
-
-  void initialise_exact(DenseMat* data, StagReal a) {
-    exact_kde = stag::ExactGaussianKDE(data, a, min_idx, max_idx + 1);
-  }
-
-  StagReal estimate_weight(const stag::DataPoint& q, const StagInt q_id) {
-    StagReal weight;
-    cache_mutex.lock();
-    if (!cached_weights.contains(q_id)) {
-      cache_mutex.unlock();
-      if (!below_cutoff) {
-        if (depth > depth_cutoff) {
-          // If we are below the cutoff depth, query the estimator.
-          weight = (StagReal) n_node * this_estimator.query(q);
-        } else {
-          // If we are above the cutoff depth, get the estimate from the
-          // children.
-          weight = left_child->estimate_weight(q, q_id) + right_child->estimate_weight(q, q_id);
-        }
-      } else {
-        weight = (StagReal) n_node * exact_kde.query(q);
-      }
-
-      cache_mutex.lock();
-      cached_weights[q_id] = weight;
-      cache_mutex.unlock();
-
-    } else {
-      weight = cached_weights[q_id];
-      cache_mutex.unlock();
-    }
-    return weight;
-  }
-
-  std::vector<StagReal> estimate_weights(DenseMat* q, indicators::ProgressBar* prog_bar,
-                                         StagInt progress_start, StagInt progress_end) {
-    std::vector<StagReal> weights;
-    if (!below_cutoff) {
-      if (depth > depth_cutoff) {
-        weights = this_estimator.query(q);
-
-        for (auto & weight : weights) {
-          weight = (StagReal) n_node * weight;
-        }
-      } else {
-        auto progress_middle = (StagInt) (progress_start + progress_end) / 2;
-        std::vector<StagReal> left_weights = left_child->estimate_weights(q, prog_bar, progress_start, progress_middle);
-        std::vector<StagReal> right_weights = right_child->estimate_weights(q, prog_bar, progress_middle, progress_end);
-        assert(left_weights.size() == right_weights.size());
-        for (StagInt i = 0; i < (StagInt) left_weights.size(); i++) {
-          weights.push_back(left_weights.at(i) + right_weights.at(i));
-        }
-      }
-    } else {
-      weights = exact_kde.query(q);
-
-      for (auto & weight : weights) {
-        weight = (StagReal) n_node * weight;
-      }
-    }
-
-    for (auto q_id = 0; q_id < (StagInt) weights.size(); q_id++) {
-      cached_weights[q_id] = weights.at(q_id);
-    }
-
-    prog_bar->set_progress(progress_end);
-
-    return weights;
-  }
-
-  std::vector<StagInt> sample_neighbors(const stag::DataPoint& q, const StagInt q_id, StagInt num_to_sample) {
-    if (below_cutoff) {
-      std::vector<StagReal> rs;
-      for (auto i = 0; i < num_to_sample; i++) {
-        rs.push_back(sampling_dist(*stag::get_global_rng()));
-      }
-      StagReal deg = estimate_weight(q, q_id); // should be cached
-      return exact_kde.sample_neighbors(q, deg, rs);
-    } else {
-      StagReal left_est = left_child->estimate_weight(q, q_id);
-      StagReal right_est = right_child->estimate_weight(q, q_id);
-      StagReal my_est = left_est + right_est;
-
-      // Get the number of left samples from a binomial distribution.
-      std::binomial_distribution<> d(num_to_sample, left_est / my_est);
-      StagInt num_left_samples = d(*stag::get_global_rng());
-
-      assert(num_left_samples >= 0);
-      std::vector<StagInt> left_samples;
-      std::vector<StagInt> right_samples;
-      if (num_left_samples > 0) left_samples = left_child->sample_neighbors(q, q_id, num_left_samples);
-      if (num_left_samples < num_to_sample) right_samples = right_child->sample_neighbors(q, q_id, num_to_sample - num_left_samples);
-      assert((StagInt) left_samples.size() + (StagInt) right_samples.size() == num_to_sample);
-
-      std::vector<StagInt> samples;
-      samples.reserve(num_to_sample);
-      samples.insert(samples.end(), left_samples.begin(), left_samples.end());
-      samples.insert(samples.end(), right_samples.begin(), right_samples.end());
-      return samples;
-    }
-  }
-
-private:
-  bool below_cutoff;
-  stag::CKNSGaussianKDE this_estimator;
-  stag::ExactGaussianKDE exact_kde;
-  StagInt min_idx;
-  StagInt max_idx;
-  StagInt n_node;
-  KDETreeEntry* left_child;
-  KDETreeEntry* right_child;
-  std::uniform_real_distribution<double> sampling_dist;
-  std::unordered_map<StagInt, StagReal> cached_weights;
-  std::mutex cache_mutex;
-  StagInt depth;
-  StagInt depth_cutoff;
-};
-
-void sample_asg_edges(DenseMat* data,
-                      KDETreeEntry& tree_root,
-                      StagInt chunk_start,
-                      StagInt chunk_end,
-                      std::vector<EdgeTriplet>& edges,
-                      StagInt edges_per_node,
-                      std::vector<StagReal>& degrees,
-                      indicators::ProgressBar* prog_bar) {
-  auto nodes_per_tick = (StagInt) ceil((StagReal) data->rows() / 70);
-  std::uniform_real_distribution<double> sampling_dist(0, 0.8);
-  auto progress_offset = (StagInt) (sampling_dist(*stag::get_global_rng()) * nodes_per_tick);
-  auto total_ticks = (StagInt) (chunk_end - chunk_start) / nodes_per_tick;
-  StagInt ticks_made = 0;
-
-  StagInt next_index = 2 * (chunk_start * edges_per_node);
-  for (auto i = chunk_start; i < chunk_end; i++) {
-    stag::DataPoint q(*data, i);
-
-    std::vector<StagInt> node_samples = tree_root.sample_neighbors(q, i, edges_per_node);
-
-    assert(i <= (StagInt) degrees.size());
-    StagReal added_weight = degrees[i] / 5;
-
-    for (auto j = 0; j < edges_per_node; j++) {
-      assert(j < (StagInt) node_samples.size());
-      StagInt neighbor = node_samples[j];
-      assert(next_index < (StagInt) edges.size());
-      edges[next_index] = EdgeTriplet(i, neighbor, added_weight);
-      next_index++;
-      edges[next_index] = EdgeTriplet(neighbor, i, added_weight);
-      next_index++;
-    }
-
-    // Update the progress bar
-    if ((i - chunk_start + progress_offset) % nodes_per_tick == 0) {
-      if (ticks_made < total_ticks) {
-        prog_bar->tick();
-        ticks_made++;
-      }
-    }
-  }
-}
-
-stag::Graph stag::approximate_similarity_graph(DenseMat* data, StagReal a) {
-  // Work out how many total edges we will sample
-  auto edges_per_node = (StagInt) (3 * log((StagReal) data->rows()));
-  StagInt num_edges = data->rows() * edges_per_node;
-
-  // Track the construction with a progress bar
-  indicators::ProgressBar prog_bar{indicators::option::BarWidth{50},
-                                   indicators::option::Start{"["},
-                                   indicators::option::Fill{"■"},
-                                   indicators::option::Lead{"■"},
-                                   indicators::option::Remainder{"-"},
-                                   indicators::option::End{"]"},
-                                   indicators::option::ForegroundColor{indicators::Color::cyan},
-                                   indicators::option::FontStyles{std::vector<indicators::FontStyle>{indicators::FontStyle::bold}},
-                                   indicators::option::PostfixText("Initialising KDE Estimators")};
-  prog_bar.set_progress(1);
-
-  // Begin by creating the tree of kernel density estimators.
-  // Creating each node is parallelized by the KDE code.
-  KDETreeEntry tree_root(data, a, 0, data->rows() - 1, 0, (StagInt) log((StagReal) edges_per_node));
-
-  prog_bar.set_option(indicators::option::PostfixText("Estimating node degrees"));
-  prog_bar.set_progress(10);
-
-  // First, compute the degrees of all nodes
-  std::vector<StagReal> degrees = tree_root.estimate_weights(data, &prog_bar, 10, 30);
-
-  prog_bar.set_option(indicators::option::PostfixText("Sampling edges"));
-  prog_bar.set_progress(30);
-
-  StagInt num_threads = std::thread::hardware_concurrency();
-  std::vector<EdgeTriplet> graph_edges(2 * num_edges);
-  if (data->rows() <= num_threads * 2) {
-    sample_asg_edges(data, tree_root, 0, data->rows(), graph_edges, edges_per_node, degrees, &prog_bar);
-  } else {
-    // Start the thread pool
-    ctpl::thread_pool pool((int) num_threads);
-    std::vector<std::future<void>> futures;
-
-    StagInt chunk_size = floor((StagReal) data->rows() / (StagReal) num_threads);
-
-    for (auto chunk_id = 0; chunk_id < num_threads; chunk_id++) {
-      StagInt this_chunk_start = chunk_id * chunk_size;
-      StagInt this_chunk_end = this_chunk_start + chunk_size;
-      if (chunk_id == num_threads - 1) this_chunk_end = data->rows();
-
-      assert(this_chunk_start <= (StagInt) data->rows());
-      assert(this_chunk_end <= (StagInt) data->rows());
-      assert(this_chunk_end > this_chunk_start);
-
-      futures.push_back(
-        pool.push(
-          [&, this_chunk_start, this_chunk_end, edges_per_node] (int id) {
-            ignore_warning(id);
-            sample_asg_edges(data, tree_root, this_chunk_start,
-                             this_chunk_end, graph_edges, edges_per_node, degrees, &prog_bar);
-          }
-        )
-      );
-    }
-
-    // Join the futures
-    for (auto chunk_id = 0; chunk_id < num_threads; chunk_id++) {
-      futures[chunk_id].get();
-    }
-  }
-
-  prog_bar.set_option(indicators::option::PostfixText("Done!"));
-  prog_bar.set_progress(100);
-
-  // Return a graph
-  SprsMat adj_mat(data->rows(), data->rows());
-  adj_mat.setFromTriplets(graph_edges.begin(), graph_edges.end());
-  return stag::Graph(adj_mat);
-}
-
-stag::Graph stag::similarity_graph(DenseMat* data, StagReal a) {
-  StagInt n = data->rows();
-  std::vector<EdgeTriplet> graph_edges;
-
-  std::vector<stag::DataPoint> datapoints = stag::matrix_to_datapoints(data);
-
-  StagInt num_threads = std::thread::hardware_concurrency();
-  if (n <= num_threads * 2) {
-    for (StagInt i = 0; i < n; i++) {
-      for (StagInt j = i; j < n; j++) {
-        StagReal weight = stag::gaussian_kernel(a, datapoints.at(i), datapoints.at(j));
-        EdgeTriplet e1(i, j, weight);
-        graph_edges.push_back(e1);
-        if (i != j) {
-          EdgeTriplet e2(j, i, weight);
-          graph_edges.push_back(e2);
-        }
-      }
-    }
-  } else {
-    // Start the thread pool
-    ctpl::thread_pool pool((int) num_threads);
-    std::vector<std::future<std::vector<EdgeTriplet>>> futures;
-
-    StagInt chunk_size = floor((StagReal) n / (StagReal) num_threads);
-
-    for (auto chunk_id = 0; chunk_id < num_threads; chunk_id++) {
-      StagInt this_chunk_start = chunk_id * chunk_size;
-      StagInt this_chunk_end = this_chunk_start + chunk_size;
-      if (chunk_id == num_threads - 1) this_chunk_end = n;
-
-      assert(this_chunk_start <= n);
-      assert(this_chunk_end <= n);
-      assert(this_chunk_end > this_chunk_start);
-
-      futures.push_back(
-          pool.push(
-              [&, this_chunk_start, this_chunk_end] (int id) {
-                ignore_warning(id);
-                std::vector<EdgeTriplet> this_thread_edges;
-                for (StagInt i = this_chunk_start; i < this_chunk_end; i++) {
-                  for (StagInt j = i; j < n; j++) {
-                    StagReal weight = stag::gaussian_kernel(a, datapoints.at(i), datapoints.at(j));
-                    EdgeTriplet e1(i, j, weight);
-                    this_thread_edges.push_back(e1);
-                    if (i != j) {
-                      EdgeTriplet e2(j, i, weight);
-                      this_thread_edges.push_back(e2);
-                    }
-                  }
-                }
-                return this_thread_edges;
-              }
-          )
-      );
-    }
-
-    // Join the futures
-    for (auto chunk_id = 0; chunk_id < num_threads; chunk_id++) {
-      std::vector<EdgeTriplet> this_thread_edges = futures[chunk_id].get();
-      graph_edges.insert(graph_edges.end(), this_thread_edges.begin(), this_thread_edges.end());
-    }
-
-    assert((StagInt) graph_edges.size() == n * n);
-  }
-
-  // Return a graph
-  SprsMat adj_mat(n, n);
-  adj_mat.setFromTriplets(graph_edges.begin(), graph_edges.end());
-  return stag::Graph(adj_mat);
 }
