+++ stag_lib/KDE/data.h
@@ -99,22 +99,6 @@ namespace stag{
    * @return a vector of stag::DataPoint pointing to the rows of the matrix
    */
   std::vector<stag::DataPoint> matrix_to_datapoints(DenseMat* data);
-
-  /**
-   * TODO: Write description.
-   * @param data
-   * @param a
-   * @return
-   */
-  Graph approximate_similarity_graph(DenseMat* data, StagReal a);
-
-  /**
-   * TODO: write description.
-   * @param data
-   * @param a
-   * @return
-   */
-  Graph similarity_graph(DenseMat* data, StagReal a);
 }
 
 #endif //STAG_LIBRARY_DATA_H
