+++ test/data_test.cpp
@@ -110,64 +110,3 @@ TEST(DataTest, MatrixToDataPoints) {
     }
   }
 }
-
-TEST(DataTest, ASGMNIST) {
-  // Load the MNIST dataset
-  std::string filename = "test/data/mnist.txt";
-  DenseMat data = stag::load_matrix(filename);
-  StagReal a = 0.000001;
-
-  // Create tha approximate similarity graph from this matrix.
-  stag::Graph asg = stag::approximate_similarity_graph(&data, a);
-  EXPECT_EQ(asg.number_of_vertices(), data.rows());
-
-  // Load the correct clusters
-  std::string labels_filename = "test/data/mnist_labels.txt";
-  DenseMat labels = stag::load_matrix(labels_filename);
-  std::vector<StagInt> labels_vec(labels.rows());
-  for (auto i = 0; i < labels.rows(); i++) {
-    labels_vec.at(i) = (StagInt) labels(i, 0);
-  }
-
-  // Check the clustering performance with the asg roughly matches the
-  // performance with the fully connected graph
-  StagInt k = 10;
-  std::vector<StagInt> clusters = stag::spectral_cluster(&asg, k);
-  StagReal asg_ari = stag::adjusted_rand_index(clusters, labels_vec);
-
-  stag::Graph sg = stag::similarity_graph(&data, a);
-  clusters = stag::spectral_cluster(&sg, k);
-  StagReal fc_ari = stag::adjusted_rand_index(clusters, labels_vec);
-  EXPECT_GE(asg_ari, 0.8 * fc_ari);
-}
-
-TEST(DataTest, ASGmoons) {
-  // Load the moons dataset
-  std::string filename = "test/data/moons.txt";
-  DenseMat data = stag::load_matrix(filename);
-  StagReal a = 10;
-
-  // Create tha approximate similarity graph from this matrix.
-  stag::Graph asg = stag::approximate_similarity_graph(&data, a);
-  EXPECT_EQ(asg.number_of_vertices(), data.rows());
-
-  // Load the correct clusters
-  std::string labels_filename = "test/data/moons_labels.txt";
-  DenseMat labels = stag::load_matrix(labels_filename);
-  std::vector<StagInt> labels_vec(labels.rows());
-  for (auto i = 0; i < labels.rows(); i++) {
-    labels_vec.at(i) = (StagInt) labels(i, 0);
-  }
-
-  // Check the clustering performance with the asg roughly matches the
-  // performance with the fully connected graph
-  StagInt k = 2;
-  std::vector<StagInt> clusters = stag::spectral_cluster(&asg, k);
-  StagReal asg_ari = stag::adjusted_rand_index(clusters, labels_vec);
-
-  stag::Graph sg = stag::similarity_graph(&data, a);
-  clusters = stag::spectral_cluster(&sg, k);
-  StagReal fc_ari = stag::adjusted_rand_index(clusters, labels_vec);
-  EXPECT_GE(fc_ari, 0.9);
-  EXPECT_GE(asg_ari, 0.8 * fc_ari);
-}
